+++ mcp-config.json
@@ -4,13 +4,13 @@
       "command": "curl",
       "args": [
         "-X", "POST",
-        "http://localhost:3000/api/v1/mcp",
+        "http://localhost:3000/mcp/v1/mcp",
         "-H", "Content-Type: application/json",
         "-d", "@-"
       ],
       "transport": {
         "type": "http",
-        "url": "http://localhost:3000/api/v1/mcp"
+        "url": "http://localhost:3000/mcp/v1/mcp"
       }
     }
   }
