+++ src/index.js
@@ -1,6 +1,6 @@
 import process from 'node:process'
 
-import { startServer } from './api/server.js'
+import { startServer } from './mcp/server.js'
 import { createLogger } from './common/logging/logger.js'
 import { setupProxy } from './common/proxy/setup-proxy.js'
