+++ src/mcp/plugins/pulse.js
