+++ src/mcp/plugins/request-logger.js
