+++ src/mcp/probes/health/router.js
