+++ src/mcp/server.js
