+++ src/mcp/v1/notes/services/mcp-notes.js
