+++ test/integration/narrow/mcp/probes/health.test.js
