+++ test/unit/mcp/server.test.js
